pyproject.toml
[tool.poetry]
name = "smart-reconciler"
version = "0.1.0"
description = ""
authors = ["LeChristopher J Blackwell <Codeblackwell@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
duckdb = "^1.3.1"
pandas = "^2.3.0"
pyarrow = "^20.0.0"
requests = "^2.32.4"
tqdm = "^4.67.1"
sentence-transformers = "^4.1.0"
geopy = "^2.4.1"
shapely = "^2.1.1"
fastapi = "^0.115.13"
uvicorn = "^0.34.3"
pydantic = "^2.11.7"
geopandas = "^1.1.0"
python-dotenv = "^1.1.0"

[tool.poetry.dev-dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
